Move fullcalendar from gem to webpack
