+++ app/assets/javascripts/application.js
@@ -21,17 +21,8 @@
 //= require adminlte.min
 //= require cocoon
 //= require toastr
-//= require fullcalendar
 //= require quagga
 //= require_tree .
-
-function isMobileResolution() {
-  return $(window).width() < 992;
-}
-
-function isShortHeightScreen() {
-  return $(window).height() < 768 && !isMobileResolution();
-}
 
 window.setTimeout(function() {
   // When the user is given an error message, we should not auto-hide it so that
@@ -45,20 +36,6 @@ window.setTimeout(function() {
 toastr.options = {
   "timeOut": "1400"
 }
-
-$(document).ready(function () {
-  const isMobile = isMobileResolution();
-  const isShortHeight = isShortHeightScreen();
-
-  $('#calendar').fullCalendar({
-    firstDay: 1,
-    displayEventTime: true,
-    eventLimit: true,
-    events: 'schedule.json',
-    height: isMobile || isShortHeight ? 'auto' : 'parent',
-    defaultView: isMobile ? 'listWeek' : 'month'
-  });
-});
 
 $(document).ready(function() {
   Filterrific.init();
