+++ app/assets/stylesheets/application.scss
@@ -11,7 +11,6 @@
  * It is generally better to create a new file per style scope.
  *
  *= require toastr
- *= require fullcalendar
  *= require barcode_reader
  */
