+++ package.json
@@ -11,6 +11,10 @@
     "@babel/preset-env": "^7.16.11",
     "@fortawesome/fontawesome-free": "^5.11.2",
     "@fortawesome/free-solid-svg-icons": "^5.11.2",
+    "@fullcalendar/core": "^5.11.0",
+    "@fullcalendar/daygrid": "^5.11.0",
+    "@fullcalendar/list": "^5.11.0",
+    "@fullcalendar/luxon": "^5.11.0",
     "@rails/actiontext": "^7.0.2",
     "@rails/webpacker": "5.4.3",
     "actiontext": "https://github.com/rails/actiontext#archive",
